Deltatime lets go

+++ Game/Source/Render.cpp
@@ -76,8 +76,8 @@ bool Render::Update(float dt)
 
 	if (app->scene->gameplayState == Scene::GameplayState::PLAYING)
 	{
-		camera.x = -app->player->position.x*scale + camera.w/2;
-		camera.y = -app->player->position.y*scale + camera.h/2;
+		camera.x = -(int)app->player->position.x * scale + camera.w / 2;
+		camera.y = -(int)app->player->position.y * scale + camera.h / 2;
 
 		//Camera limits ---> HARDCODED
 		if (camera.x > -160)
